proyecto comandas

+++ dtos/src/main/java/dto/IngredienteSeleccionadoDTO.java
@@ -12,12 +12,12 @@ public class IngredienteSeleccionadoDTO {
 
     private Long id;
     private String nombre;
-    private Integer cantidad;
+    private Double cantidad;
 
     public IngredienteSeleccionadoDTO() {
     }
 
-    public IngredienteSeleccionadoDTO(Long id, String nombre, Integer cantidad) {
+    public IngredienteSeleccionadoDTO(Long id, String nombre, Double cantidad) {
         this.id = id;
         this.nombre = nombre;
         this.cantidad = cantidad;
@@ -39,11 +39,11 @@ public class IngredienteSeleccionadoDTO {
         this.nombre = nombre;
     }
 
-    public Integer getCantidad() {
+    public Double getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(Integer cantidad) {
+    public void setCantidad(Double cantidad) {
         this.cantidad = cantidad;
     }
